+++ cli/cli/kurtosis_config/resolved_config/kurtosis_cluster_config_test.go
@@ -8,21 +8,21 @@ import (
 
 func TestNewKurtosisClusterConfigEmptyOverrides(t *testing.T) {
 	kurtosisClusterConfigOverrides := v2.KurtosisClusterConfigV2{}
-	_, err := NewKurtosisClusterConfigFromOverrides(&kurtosisClusterConfigOverrides)
+	_, err := NewKurtosisClusterConfigFromOverrides("test", &kurtosisClusterConfigOverrides)
 	require.Error(t, err)
 }
 
 func TestNewKurtosisClusterConfigDockerType(t *testing.T) {
 	dockerType := KurtosisClusterType_Docker.String()
 	kurtosisClusterConfigOverrides := v2.KurtosisClusterConfigV2{Type: &dockerType}
-	_, err := NewKurtosisClusterConfigFromOverrides(&kurtosisClusterConfigOverrides)
+	_, err := NewKurtosisClusterConfigFromOverrides("test", &kurtosisClusterConfigOverrides)
 	require.NoError(t, err)
 }
 
 func TestNewKurtosisClusterConfigKubernetesNoConfig(t *testing.T) {
 	kubernetesType := KurtosisClusterType_Kubernetes.String()
 	kurtosisClusterConfigOverrides := v2.KurtosisClusterConfigV2{Type: &kubernetesType}
-	_, err := NewKurtosisClusterConfigFromOverrides(&kurtosisClusterConfigOverrides)
+	_, err := NewKurtosisClusterConfigFromOverrides("test", &kurtosisClusterConfigOverrides)
 	require.Error(t, err)
 }
 
@@ -31,7 +31,7 @@ func TestNewKurtosisClusterConfigNonsenseType(t *testing.T) {
 	kurtosisClusterConfigOverrides := v2.KurtosisClusterConfigV2{
 			Type: &clusterType,
 	}
-	_, err := NewKurtosisClusterConfigFromOverrides(&kurtosisClusterConfigOverrides)
+	_, err := NewKurtosisClusterConfigFromOverrides("test", &kurtosisClusterConfigOverrides)
 	require.Error(t, err)
 }
 
@@ -45,7 +45,7 @@ func TestNewKurtosisClusterConfigKubernetesPartialConfig(t *testing.T) {
 		Type: &kubernetesType,
 		Config: &kubernetesPartialConfig,
 	}
-	_, err := NewKurtosisClusterConfigFromOverrides(&kurtosisClusterConfigOverrides)
+	_, err := NewKurtosisClusterConfigFromOverrides("test", &kurtosisClusterConfigOverrides)
 	require.Error(t, err)
 }
 
@@ -63,6 +63,6 @@ func TestNewKurtosisClusterConfigKubernetesFullConfig(t *testing.T) {
 		Type: &kubernetesType,
 		Config: &kubernetesFullConfig,
 	}
-	_, err := NewKurtosisClusterConfigFromOverrides(&kurtosisClusterConfigOverrides)
+	_, err := NewKurtosisClusterConfigFromOverrides("test", &kurtosisClusterConfigOverrides)
 	require.NoError(t, err)
 }