+++ cli/docs/changelog.md
@@ -4,6 +4,8 @@
 ### Fixes
 * Fixes bugs that were happening with API container's gRPC wire format
 * Fixes various bugs with files artifacts expansion
+* Fixed `RemoveService` in APIC gateway not killing local connection
+* Fixed APIC gateway trying to create a connection to a Kurtosis service with no ports specified
 
 ### Breaking Changes
 * Upgraded to engine 1.25.3